requirements.txt
Flask==3.1.0
gunicorn==21.2.0
line-bot-sdk==3.5.0
openai==1.3.0
stripe==10.12.0
bcrypt==4.1.2
PyPDF2==3.0.1
python-docx==1.0.1
requests==2.31.0
Jinja2==3.1.2
Werkzeug==3.1.3
APScheduler==3.10.4
pytz==2024.1
Flask-Limiter==3.5.0
PyJWT==2.8.0
reportlab==4.0.9
qrcode==7.4.2
chardet==5.2.0
pillow==10.2.0
google-cloud-firestore==2.19.0
supabase==2.10.0
